+++ karma.config.js
@@ -1,80 +1,43 @@
 // Karma configuration for LaxarJS core
 /* eslint-env node */
 
-
 const webpackConfig = Object.assign( {}, require('./webpack.base.config' ) );
 delete webpackConfig.entry.laxar;
 delete webpackConfig.plugins;
 webpackConfig.devtool = 'inline-source-map';
-webpackConfig.node = { fs: 'empty' };
 
 module.exports = function(config) {
-   config.set({
+   const browsers = [ 'PhantomJS', 'Firefox' ].concat( [
+      process.env.TRAVIS ? 'ChromeTravisCi' : 'Chrome'
+   ] );
 
-      // base path that will be used to resolve all patterns (eg. files, exclude)
-      basePath: '',
-
-      // frameworks to use
-      // available frameworks: https://npmjs.org/browse/keyword/karma-adapter
+   config.set( {
       frameworks: [ 'jasmine' ],
-
-      // list of files / patterns to load in the browser
       files: [
          'polyfills.js',
          'lib/*/spec/spec-runner.js'
       ],
-
-      // list of files to exclude
-      exclude: [
-      ],
-
-      // preprocess matching files before serving them to the browser
-      // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
       preprocessors: {
          'polyfills.js':  [ 'webpack', 'sourcemap' ],
          'lib/*/spec/spec-runner.js': [ 'webpack', 'sourcemap' ]
       },
       webpack: webpackConfig,
 
-      // test results reporter to use
-      // possible values: 'dots', 'progress'
-      // available reporters: https://npmjs.org/browse/keyword/karma-reporter
-      reporters: [ 'progress' ],
-      unit: {
-         singleRun: true,
-      },
+      reporters: [ 'progress', 'junit' ],
       junitReporter: {
          outputDir: 'karma-output/'
       },
-
-      // web server port
       port: 9876,
-
-      // enable / disable colors in the output (reporters and logs)
-      colors: true,
-
-      // level of logging
-      // possible values: config.LOG_DISABLE || config.LOG_ERROR || config.LOG_WARN || config.LOG_INFO || config.LOG_DEBUG
-      logLevel: config.LOG_INFO,
-
-      // start these browsers
-      // available browser launchers: https://npmjs.org/browse/keyword/karma-launcher
-      browsers: [
-         // 'MSIE',
-         'PhantomJS',
-         // 'Chrome',
-         // 'Firefox'
-      ],
+      browsers: browsers,
+      customLaunchers: {
+         ChromeTravisCi: {
+            base: 'Chrome',
+            flags: [ '--no-sandbox' ]
+         }
+      },
       browserNoActivityTimeout: 100000,
-
-      // Continuous Integration mode
-      // if true, Karma captures browsers, runs the tests and exits
       singleRun: true,
-      // enable / disable watching file and executing tests whenever any file changes
       autoWatch: false,
-
-      // Concurrency level
-      // how many browser should be started simultaneous
       concurrency: Infinity
-   });
+   } );
 };